0-read_textfile.c

+++ 0x15-file_io/0-read_textfile.c
@@ -8,12 +8,12 @@
  * Return: w- actual number of bytes read and printed
  * 0 when function fails or filename is NULL.
  */
-int read_textfile(const char *filename, size_t letters)
+ssize_t read_textfile(const char *filename, size_t letters)
 {
 char *buf;
-ssize_t fd;
+int fd;
 ssize_t w;
-ssize_t t;
+int t;
 
 fd = open(filename, O_RDONLY);
 if (fd == -1)
